Added toggle to camera scaling
In the camera3d node, you can toggle whether or not to have the camera's distance from the ball scale with the size of the ball. You can also adjust the intensity of this scaling.

I have it set to "false" by default since it messed up the smoothed out motion of the camera.