Imported snowball model & made jump check for floor
I imported a snowball model, but wasn't able to replace the current ball model without messing up how the ball rolled.

I also made the ball check for the floor before jumping, but I believe the ball ends up scaling up so much that the floor check is within the ball and it never thinks the ball is touching the ground.